agents_catalog/26-Medical-Device-Agent/requirements.txt
streamlit==1.45.1
boto3==1.38.18

strands-agents==0.1.2
strands-agents-tools>=0.1.1
requests>=2.32.4
python-dotenv>=1.0.0